snowflake-manifest-from-share-library/requirements.txt
snowflake-connector-python>=2.7.0
PyYAML>=6.0